cp before refactor location of cdk libs
