--- v2/ssm_agent/cdk/lib/IotPolicy/assets/requirements.txt
@@ -1,1 +1,0 @@
-cryptography==3.4.7
